pyproject.toml
[project]
name = "gticl"
version = "0.1.0"
description = "Implementation of the GTICL and TICL algorithm"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.3.0",
    "litellm>=1.61.20",
    "pip>=25.0.1",
    "pydantic>=2.10.6",
    "tabulate>=0.9.0",
]
